==> Codes/1. Filter Asian and European Countries.ipynb <==
import numpy as np
import pandas as pd

# 1. Data

### 1.1. OXCGRT Data 

data = pd.read_csv('../Data/Covid19_All_20210710.csv', low_memory=False, index_col=0)

### 1.2. Continent Data

continent = pd.read_csv('../Data/continent_data.csv', low_memory=False, index_col=0)
continent = continent[['Continent_Name', 'Three_Letter_Country_Code']]
continent.columns = ['ContinentName', 'CountryCode']

# add Kosovo
continent = continent.append({'ContinentName': 'Europe', 
                              'CountryCode': 'RKS'}, ignore_index=True)

# 2. Merge

data_merged = data.merge(continent, how='left', on='CountryCode')

# 3. Filtering

cols = ['CountryName', 'ContinentName', 'RegionName', 'Date', 'C1_School closing',
       'C2_Workplace closing', 'C3_Cancel public events', 
       'C4_Restrictions on gatherings', 'C5_Close public transport',
       'C6_Stay at home requirements', 'C7_Restrictions on internal movement', 
       'C8_International travel controls', 'E1_Income support',
       'E2_Debt/contract relief', 'E3_Fiscal measures',
       'E4_International support', 'H1_Public information campaigns', 
       'H2_Testing policy', 'H3_Contact tracing', 'H4_Emergency investment in healthcare', 
       'H5_Investment in vaccines', 'H6_Facial Coverings', 
       'H7_Vaccination policy', 'H8_Protection of elderly people', 
       'ConfirmedCases', 'ConfirmedDeaths']

### Asia

asia = data_merged.loc[data_merged.ContinentName == 'Asia', cols]

asia.sample(5)

asia.to_csv('Asia_covid19_20210710.csv')

### Europe

europe = data_merged.loc[data_merged.ContinentName == 'Europe', cols]

europe.sample(5)

europe.to_csv('Europe_covid19_20210710.csv')

### USA

usa = data_merged.loc[data_merged.CountryName == 'United States']

usa.sample(5)

usa.to_csv('../Data/Covid19_USA.csv')
