==> Codes/2. Vaccination Data (Countries).ipynb <==
import numpy as np
import pandas as pd

# Get Data 

https://github.com/owid/covid-19-data/tree/master/public/data/vaccinations

**daily_vaccinations**: new doses administered per day (7-day smoothed). For countries that don't report data on a daily basis, we assume that doses changed equally on a daily basis over any periods in which no data was reported. This produces a complete series of daily figures, which is then averaged over a rolling 7-day window. An example of how we perform this calculation can be found here.

**people_vaccinated**: total number of people who received at least one vaccine dose. If a person receives the first dose of a 2-dose vaccine, this metric goes up by 1. If they receive the second dose, the metric stays the same.

**people_fully_vaccinated**: total number of people who received all doses prescribed by the vaccination protocol. If a person receives the first dose of a 2-dose vaccine, this metric stays the same. If they receive the second dose, the metric goes up by 1.

This dataset includes some subnational locations (England, Northern Ireland, Scotland, Wales, Northern Cyprus…) and international aggregates (World, continents, European Union…). They can be identified by their iso_code that starts with OWID_.

url = 'https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/vaccinations/vaccinations.csv'

vaccination = pd.read_csv(url)

### Filter Features

vaccination.columns

cols = ['location', 'iso_code', 'date', 'people_vaccinated', 'people_fully_vaccinated', 'daily_vaccinations']

vaccination = vaccination[cols]

vaccination.sample(5)

### Filter Countries

countries = ['Sweden', 'France']

vaccination.loc[vaccination.location == 'United States']

### Save

vaccination.to_csv('../Data/Vaccination.csv')
