==> Codes/2. Vaccination Data (States).ipynb <==
import numpy as np
import pandas as pd

# Get Data 

vaccination = pd.read_csv('../Data/US_State_Vaccinations.csv')
vaccination.head()

### Filter Features

vaccination.columns

cols = ['date', 'location', 'daily_vaccinations', 'daily_vaccinations_per_million', 
        'people_vaccinated_per_hundred', 'people_fully_vaccinated_per_hundred' ]

vaccination = vaccination[cols]

vaccination.sample(5)

### Filter Countries

### Save

vaccination.to_csv('../Data/Vaccination_USA_Filtered.csv')
