==> Codes/3. Weather Dataset(NOAA).ipynb <==
import numpy as np
import pandas as pd

import requests
import datetime

import os
import sys

Check ou this website for further documentaion regarding the weather API. <br>
https://www.ncdc.noaa.gov/cdo-web/webservices/v2#data

def get_region_info(mytoken, level='country'):
    
    """
    It gives us all the IDs for different countries, cities, states, or etc.
    """
    
    level_dict = {'country': 'CNTRY', 
                  'state': 'ST',
                  'city': 'CITY',
                  'county': 'CNTY',
                  'zip code': 'ZIP'}
    
    
    assert level in [*level_dict.keys()], f'level should be one of the {[*level_dict.keys()]}'
    
    
    token = {'token': mytoken}

    # passing as string instead of dict because NOAA API does not like percent encoding

    region = f'locationcategoryid={level_dict[level]}' + '&units=standard' + '&limit=1000'
    base_url = 'https://www.ncdc.noaa.gov/cdo-web/api/v2/locations'
    r = requests.get(base_url, headers=token, params=region)
    print("Request status code: " + str(r.status_code))

    try:
        # results comes in json form. Convert to dataframe
        df = pd.DataFrame.from_dict(r.json()['results'])
        print("Successfully retrieved " + str(len(df['id'].unique())) + " stations")

        if df.count().max() >= 1000:
            print('WARNING: Maximum data limit was reached (limit = 1000)')
            print('Consider breaking your request into smaller pieces')

        return df[['name','id']]
    # Catch all exceptions for a bad request or missing data
    except:
        print("Error converting station data to dataframe. Missing data?")
        

def get_temp(locationid, begin_date, end_date, mytoken, data='daily'):
    
    """
    Get average, minimum, and maximum temperature for a location.
    """
    
   
    data_dict = {'daily': 'GHCND',
                  'monthly': 'GSOM',
                  'yearly': 'GSOY'}
    
    
    assert data in [*data_dict.keys()], (f'data should be one of the {[*data_dict.keys()]}. For more datasets visit', 
                                        'https://www.ncdc.noaa.gov/cdo-web/api/v2/datasets')
    

    token = {'token': mytoken}

    # passing as string instead of dict because NOAA API does not like percent encoding
    params = 'datasetid=GHCND' + '&locationid=' + str(locationid) + '&startdate=' + str(begin_date) + '&enddate=' + str(
        end_date) +'&limit=1000' + '&units=standard'+'&datacategoryid=TEMP'

    base_url = 'https://www.ncdc.noaa.gov/cdo-web/api/v2/data'

    r = requests.get(base_url, params=params, headers=token)
    print("Request status code: " + str(r.status_code))
    

    try:
        # results comes in json form. Convert to dataframe
        df = pd.DataFrame.from_dict(r.json()['results'])
        print("Successfully retrieved " + str(len(df['station'].unique())) + " stations")
        dates = pd.to_datetime(df['date'])
        print("Last date retrieved: " + str(dates.iloc[-1]))

        if df.count().max() == 1000:
            print('WARNING: Maximum data limit was reached (limit = 1000)')
            print('Consider breaking your request into smaller pieces')
        
        
        tavg = df[df.datatype=='TAVG'].value.values.mean()
        tmax = df[df.datatype=='TMAX'].value.values.mean()
        tmin = df[df.datatype=='TMIN'].value.values.mean()
        prcp = df[df.datatype=='PRCP'].value.values.mean()
        
        return tavg, tmax, tmin, prcp

    # Catch all exceptions for a bad request or missing data
    except:
        print("Error converting weather data to dataframe. Missing data?")
        return np.nan, np.nan, np.nan

# Implementation

countries = ['Italy', 'France']

df_locations = get_region_info(mytoken, level='country')
df_locations_dict = df_locations.set_index('name').to_dict()['id']

WINDOW = pd.date_range(start='2020-01-01', end='2021-07-10', freq='D')

output_list = []

for country in countries:
    country_id = df_locations_dict[country]
    
    for day in WINDOW[:5]:
        
        day = day.strftime('%Y-%m-%d')
        temp = get_temp(locationid=country_id, begin_date=day, end_date=day, mytoken=mytoken, data='daily')
        temp = (*temp, country, day)

        output_list.append(temp)

        
        
output = pd.DataFrame(output_list, columns=['tavg', 'tmax', 'tmin', 'prcp', 'Country', 'Day'])

output
