==> Codes/6- Population Dataset.ipynb <==
import numpy as np
import pandas as pd

**Information**

*Data Source*: World Development Indicators <br>
*Last Update*: 6/30/2021

*Link*: https://data.worldbank.org/indicator/SP.POP.TOTL

For England:
https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/bulletins/annualmidyearpopulationestimates/mid2020

For USA:  
https://www.census.gov/programs-surveys/popest/technical-documentation/research/evaluation-estimates/2020-evaluation-estimates/2010s-state-total.html

# Data

population = pd.read_csv('../Data/Population.csv')

population.sample(10)
