==> Codes/7. Create Final Datsets.ipynb <==
import numpy as np
import pandas as pd



NPI_COLS = ['C1_School closing', 'C2_Workplace closing',
           'C3_Cancel public events', 'C4_Restrictions on gatherings',
           'C5_Close public transport', 'C6_Stay at home requirements',
           'C7_Restrictions on internal movement',
           'C8_International travel controls', 'E1_Income support',
           'E2_Debt/contract relief', 'H1_Public information campaigns',
           'H2_Testing policy', 'H3_Contact tracing', 'H6_Facial Coverings', 
           'H7_Vaccination policy','H8_Protection of elderly people']

covid_path = '../Data/Covid19_Europe_20210710.csv'
vaccination_path='../Data/Vaccination.csv'
population_path='../Data/Population.csv'
weather_path = '../Data/Weather.csv'

COUNTRIES = ['Germany', 'Italy', 'France', 'Spain',  'Denmark', 'England',
             'Netherlands', 'Belgium', 'Sweden', 'Switzerland', 'Norway'] 

### Functions

def _prepare_covid(file_path, countries_list):   
    """
    It generate the covid-19 dataset for a given list of countries, up to 2021-07-10.
    """
    covid_cols = ['CountryName', 'RegionName', 'Date', 'ConfirmedCases', 'ConfirmedDeaths']
    
    # Main Covid-19 dataset
    df = pd.read_csv(file_path, index_col=0, parse_dates=['Date']).reset_index(drop=True)
    # Only Keep needed Columns
    df = df[covid_cols + NPI_COLS]
    # Filter for countries
    df_filtered = df.loc[df.CountryName.isin(countries_list)]
    # Exception for England
    if 'England' in countries_list:
        # Extract only England
        england = df.loc[(df.CountryName == 'United Kingdom') & (df.RegionName == 'England')]
        # Merge countries
        df_filtered = pd.concat([df_filtered, england], axis='rows', sort=False)
        # Change United Kingdom to England
        df_filtered.loc[df_filtered.CountryName == 'United Kingdom', 'CountryName'] = 'England'   
    # Drop unnecessary columns
    df_filtered.drop(columns=['RegionName'], inplace=True)
    
    if len(df_filtered) == 0:
        raise ValueError('Check the list of countries again. The output length is zero!')

    return df_filtered

def _prepare_vaccination(file_path, countries_list):

    # Read data
    df = pd.read_csv(file_path, index_col=0, parse_dates=['date'])
    # Drop unnecessary columns
    df.drop(columns=['iso_code'], inplace=True)
    # Change column names
    df.columns = ['CountryName', 'Date', 'Vaccinated', 'FullyVaccinated', 'DailyVaccination']
    # Filter for countries
    df_filtered = df.loc[df.CountryName.isin(countries_list), :]

    return df_filtered

def _prepare_population(file_path, countries_list):
    
    df = pd.read_csv(file_path)
    # Drop unnecessary columns
    df.drop(columns=['Country Code'], inplace=True)
    # Change Columns Names
    df.columns = ['CountryName', 'Population2020']
    # Filter Countries
    df_filtered = df.loc[df.CountryName.isin(countries_list)]
    if 'England' in countries_list:
        # population of England
        df_filtered = df_filtered.append({'CountryName': 'England', 
                                          'Population2020': 56550000}, 
                                         ignore_index=True)

    return df_filtered

def _prepare_weather(file_path, countries_list):

    
    # Read Data
    df = pd.read_csv(file_path, index_col=0, parse_dates=['date_time'])
    # Change Columns Names
    df.columns = ['CountryName', 'Date', 'HeatIndexC', 'humidity', 
                  'tempC', 'windspeedKmph', 'precipMM', 'DewPointC', 'pressure']
    # Filter Countries
    df_filtered = df.loc[df.CountryName.isin(countries_list)]

    return df_filtered

def _preprocess_covid(covid_df, population_df, window_size=7):
    """
    # Fill missing values by interpolation, ffill, and filling NaN.
    # Creates new variables of interest.
    """
    df_new = covid_df.copy()
    # Interpolation & Drop country which no number of cases is available
    df_new.update(df_new.groupby('CountryName')['ConfirmedCases'].apply(lambda group: group.interpolate(limit_area='inside')))
    df_new.dropna(subset=['ConfirmedCases'], inplace=True)
    # Interpolation & Drop country which no number of cases is available
    df_new.update(df_new.groupby('CountryName')['ConfirmedDeaths'].apply(lambda group: group.interpolate(limit_area='inside')))
    df_new.dropna(subset=['ConfirmedDeaths'], inplace=True)
    # For Policies
    for npi_column in NPI_COLS:
        df_new.update(df_new.groupby('CountryName')[npi_column].ffill().fillna(0))
        
    
    # Merge covid and vaccination
    final_df = pd.merge(df_new, population_df, how='left', on='CountryName')

    # Compute number of new cases and deaths each day
    final_df['NewCases'] = final_df.groupby('CountryName').ConfirmedCases.diff().fillna(0)
    final_df['NewDeaths'] = final_df.groupby('CountryName').ConfirmedDeaths.diff().fillna(0)
    
    # Replace negative values (which do not make sense for these columns) with 0
    final_df['NewCases'] = final_df['NewCases'].clip(lower=0)
    final_df['NewDeaths'] = final_df['NewDeaths'].clip(lower=0)

    # Compute smoothed versions of new cases and deaths each day
    final_df['SmoothNewCases'] = final_df.groupby('CountryName')['NewCases'].rolling(
        window_size, center=False).mean().fillna(0).reset_index(0, drop=True).round()
    final_df['SmoothNewDeaths'] = final_df.groupby('CountryName')['NewDeaths'].rolling(
        window_size, center=False).mean().fillna(0).reset_index(0, drop=True).round()

    # Compute percent change in new cases and deaths each day
    final_df['CaseRatio'] = final_df.groupby('CountryName').SmoothNewCases.pct_change().fillna(0).replace(np.inf, 0) + 1
    final_df['DeathRatio'] = final_df.groupby('CountryName').SmoothNewDeaths.pct_change().fillna(0).replace(np.inf, 0) + 1
    
    # Add column for proportion of population infected
    final_df['ProportionInfected'] = final_df['ConfirmedCases'] / final_df['Population2020']

    # Create column of value to predict
    final_df['PredictionRatio'] = final_df['CaseRatio'] / (1 - final_df['ProportionInfected'])
    
    return final_df

def _preprocess_vaccination(df):
    
    df_new = df.copy()
    # daily vaccination: put 0 for missing values
    df_new.DailyVaccination.fillna(0, inplace=True)
    
    # FullyVaccinated: forward fill (fill with the latest value, since it is cumulative), and 0 for the rest
    df_new.update(df_new.groupby('CountryName')['FullyVaccinated'].ffill().fillna(0))
    
    # Vaccinated: first forward fill, then replace with 0 for the rest.
    df_new.update(df_new.groupby('CountryName')['Vaccinated'].ffill().fillna(0))
    
    return df_new

def load_data(covid_path, vaccination_path, population_path, weather_path, countries_list):
    
    covid_data = _prepare_covid(covid_path, countries_list)
    vaccination_data = _prepare_vaccination(vaccination_path, countries_list)
    population_data = _prepare_population(population_path, countries_list)
    weather_data = _prepare_weather(weather_path, countries_list)
    
    # Preprocessing
    vaccination_data = _preprocess_vaccination(vaccination_data)
    covid_data = _preprocess_covid(covid_data, population_data, window_size=7)
    
    # Merging Datasets
    merged_data = pd.merge(covid_data, vaccination_data, on=['CountryName', 'Date'], how='left')
    merged_data = pd.merge(merged_data, weather_data, on=['CountryName', 'Date'], how='left')

    return merged_data

### Implementation

_prepare_vaccination(vaccination_path, COUNTRIES)

_prepare_weather(weather_path, COUNTRIES)

_prepare_population(population_path, COUNTRIES)

final_df = load_data(covid_path, vaccination_path, population_path, weather_path, COUNTRIES)

final_df.head()

### Save

final_df.isna().sum()

final_df.to_csv('../Data/Covid19_Data.csv')
