==> Codes/9- Test Windows.ipynb <==
import numpy as np
import pandas as pd
from scipy.stats import pearsonr

%matplotlib inline
import matplotlib
import matplotlib.pyplot as plt
import matplotlib.dates as mdates
from matplotlib import ticker
import seaborn as sns

from plotting import plot_formatter

# Data 

data = pd.read_csv('../Data/Covid19_Data.csv', index_col=0, parse_dates=['Date'])
data= data.loc[data.CountryName.isin(['England', 'France', 'Germany', 'Spain', 'Italy'])]
data.set_index('Date', inplace=True)
data.head()

from datetime import datetime, timedelta
import pandas as pd


def generate_data(file_path, country, train_start_date, train_end_date, test_period='short'):
    
    """
    It generates the dataset that can be used for training
    
    Parameters
    ----------
    file_path: str
        path to the data file containing the preprocessed data
    country: str
        Name of teh country to use
    train_start_date: str
        The beginning period of training data (inclusive), format of yyyy-mm-dd
    train_end_date: str
        The ending date of training data (inclusive), format of yyyy-mm-dd
    test_period: str
        short (One week ahead), long (Four weeks ahead).
    
        
    Return
    ------
    A Training and test dataframe
    """
    
    # Read Data
    df = pd.read_csv(file_path, index_col=0, parse_dates=['Date'])
    df.set_index('Date', inplace=True)
    
    # Filtering Country
    df_country = df.loc[df.CountryName == country]
    
    # Filtering features
    
    
    # Training df
    df_train = df_country.loc[train_start_date:train_end_date]
    
    # Test df
    test_start_date = datetime.strptime(train_end_date, '%Y-%m-%d') + timedelta(days=1)
    if test_period == 'short':
        test_end_date = test_start_date + timedelta(weeks=1)
    elif test_period == 'long':
        test_end_date = test_start_date + timedelta(weeks=4)
        
    df_test = df_country.loc[test_start_date:test_end_date]
    
    
    return df_train

generate_data(file_path='../Data/Covid19_Data.csv',
              country='Italy', train_start_date='2020-10-01', train_end_date='2020-11-1', test_period='short').index

generate_data(file_path='../Data/Covid19_Data.csv',
              country='Italy', train_start_date='2020-10-01', train_end_date='2020-11-1', test_period='short').index

__Variants Emergence__

* Alpha $\rightarrow$ Sep-2020 
* Beta $\rightarrow$ May-2020 
* Gamma $\rightarrow$ Nov-2020 
* Delta $\rightarrow$ Oct-2020
* Lambda $\rightarrow$ Dec-2020
* Mu $\rightarrow$ Jan-2021
