==> Codes/10. ETS Models.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../Data/Sample/airline_passengers.csv')

# Smoothing Methods 

# Simple Moving Average (SMA)
pd.Series().rolling(WINDOW_SIZE).mean()

# Exponential Weighted Moving Average (EWMA)
pd.Series().ewm(alpha=ALPHA, adjust=False)

# Holt-Winters Models

**Time series frequencies:**  
https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html

### 1. Simple Exponential Smoothing 

from statsmodels.tsa.holtwinters import SimpleExpSmoothing

ses = SimpleExpSmoothing(x)
ses.fit()  # if don't specify alpha, it will optimize it for the given set

ses.predict(start, end) # giving the first and last index to predict
ses.forcast(5)          # forecast t+1, t+2, ..., t+5
ses.fittedvalues        # give the prediction for the training data

### 2. Holt's linear trend model 

from statsmodels.tsa.holtwinters import Holt

holt = Holt(x)
holt.fit()
