==> Codes/12. Tutorial_RNN Simplified.ipynb <==
from keras.layers import Dense, Input, SimpleRNN, Embedding
from keras.models import Model, Sequential

import numpy as np
import pandas as pd

N = 2
T = 10
D = 3
K = 2
X = np.random.randn(N, T, D)

M = 5

i = Input(shape=(T, D))
x = SimpleRNN(M)(i)
x = Dense(K)(x)

model = Model(i, x)

model.predict(X)

model.summary()

w_x, w_h, b_h = model.layers[1].get_weights()
w_o, b_o = model.layers[2].get_weights()

### For one obs.

x = X[0]
h_last = np.zeros(M)

for j in range(T):
    h = np.tanh(x[j].dot(w_x) + h_last.dot(w_h) + b_h)
    h_last = h
    y = h.dot(w_o) + b_o

print(y)

### For Multiple Observations

yhat = np.array([]).reshape(0, K)

for n in range(len(X)):
    x = X[n]
    h_last = np.zeros(M)
    for j in range(T):
        h = np.tanh(x[j].dot(w_x) + h_last.dot(w_h) + b_h)
        h_last = h
        y = h.dot(w_o) + b_o
    yhat = np.concatenate((yhat, np.expand_dims(y, 0)))

print(yhat)
