==> Codes/16. Data Preparation.ipynb <==
import numpy as np
import pandas as pd

NPI_COLS = ['C1_School closing', 'C2_Workplace closing',
           'C3_Cancel public events', 'C4_Restrictions on gatherings',
           'C5_Close public transport', 'C6_Stay at home requirements',
           'C7_Restrictions on internal movement',
           'C8_International travel controls', 'E1_Income support',
           'E2_Debt/contract relief', 'H1_Public information campaigns',
           'H2_Testing policy', 'H3_Contact tracing', 'H6_Facial Coverings', 
           'H7_Vaccination policy','H8_Protection of elderly people']

WINDOW = 7

# Population

https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/bulletins/annualmidyearpopulationestimates/mid2020#:~:text=1.,%25)%20since%20mid%2Dyear%202019.

pop_uk = pd.DataFrame({'CountryName': ['Wales', 'England', 'Scotland', 'Northern Ireland'],
                       'Population2020': [3170000, 56550000, 5466000, 1896000]})

# Read Data
pop = pd.read_csv('../Data/Population.csv')

# Dropping Columns
pop.drop(columns=['Country Code'], inplace=True)

# Renaming COlumns
pop.columns = ['CountryName', 'Population2020']

# Drop missing Values
pop.dropna(inplace=True)

# add UK Regions
pop = pd.concat([pop, pop_uk], axis='rows').reset_index(drop=True)

# Russia & Faroe Islands
pop.CountryName.replace(to_replace='Russian Federation',  value='Russia', inplace=True)
pop.CountryName.replace(to_replace='Faroe Islands',  value='Faeroe Islands', inplace=True)

# Change the type of Popluation column
pop['Population2020'] = pop['Population2020'].astype('int')

pop

# Vaccination

# Read
vaccine = pd.read_csv('../Data/Vaccination.csv', index_col=0, parse_dates=['date'])

# Drop Columns
vaccine.drop(columns=['iso_code'], inplace=True)

# Renaming Columns
vaccine.columns = ['CountryName', 'Date', 'Vaccinated', 'FullyVaccinated', 'DailyVaccination']

# daily vaccination: put 0 for missing values
vaccine.DailyVaccination.fillna(0, inplace=True)

# FullyVaccinated: forward fill (fill with the latest value, since it is cumulative), and 0 for the rest
vaccine.update(vaccine.groupby('CountryName')['FullyVaccinated'].ffill().fillna(0))

# Vaccinated: first forward fill, then replace with 0 for the rest.
vaccine.update(vaccine.groupby('CountryName')['Vaccinated'].ffill().fillna(0))

# Countries
vaccine.CountryName.replace(to_replace='Czechia',  value='Czech Republic', inplace=True)
vaccine.CountryName.replace(to_replace='Slovakia',  value='Slovak Republic', inplace=True)

vaccine

# Weather 

weather = pd.read_csv('../Data/Weather.csv', index_col=0, parse_dates=['date_time'])
weather.columns = ['CountryName', 'Date', 'HeatIndexC', 'humidity', 
                  'tempC', 'windspeedKmph', 'precipMM', 'DewPointC', 'pressure']

weather

# Covid

covid = pd.read_csv('../Data/Covid19_Europe_20210710.csv', 
                    index_col=0, 
                    parse_dates=['Date'])

# Filter Countries
covid = covid[['CountryName', 'RegionName', 'Date', 'ConfirmedCases', 'ConfirmedDeaths'] + NPI_COLS]

# Considering the Regions of UK as separate countries
covid = covid.loc[~((covid.CountryName == 'United Kingdom') & (covid.RegionName.isna()))]

is_england = ((covid.CountryName == 'United Kingdom') & (covid.RegionName == 'England')).values
is_ireland = ((covid.CountryName == 'United Kingdom') & (covid.RegionName == 'Northern Ireland')).values
is_scotland = ((covid.CountryName == 'United Kingdom') & (covid.RegionName == 'Scotland')).values
is_wales = ((covid.CountryName == 'United Kingdom') & (covid.RegionName == 'Wales')).values

covid.loc[is_england, 'CountryName'] = 'England'
covid.loc[is_ireland, 'CountryName'] = 'Northern Ireland'
covid.loc[is_scotland, 'CountryName'] = 'Scotland'
covid.loc[is_wales, 'CountryName'] = 'Wales'

# Removing Region Name
covid.drop(columns=['RegionName'], inplace=True)

# Interpolation & Drop country which no number of cases is available
covid['ConfirmedCases'] = covid.groupby('CountryName')['ConfirmedCases'].\
                                apply(lambda group: group.interpolate(limit_area='inside').ffill().fillna(0))

# Interpolation & Drop country which no number of deaths is available
covid['ConfirmedDeaths'] = covid.groupby('CountryName')['ConfirmedDeaths'].\
                                apply(lambda group: group.interpolate(limit_area='inside').ffill().fillna(0))


# Fill missing values for Policies
for npi_column in NPI_COLS:
    covid[npi_column] = covid.groupby('CountryName')[npi_column].ffill().fillna(0).astype('int')

# Merging

# Merge covid and population
final_df = pd.merge(covid, pop, how='left', on='CountryName')

# Compute number of new cases and deaths each day
final_df['NewCases'] = final_df.groupby('CountryName').ConfirmedCases.diff().fillna(0)
final_df['NewDeaths'] = final_df.groupby('CountryName').ConfirmedDeaths.diff().fillna(0)

# Replace negative values (which do not make sense for these columns) with 0
final_df['NewCases'] = final_df['NewCases'].clip(lower=0)
final_df['NewDeaths'] = final_df['NewDeaths'].clip(lower=0)

# Compute smoothed versions of new cases and deaths each day
final_df['SmoothNewCases'] = final_df.groupby('CountryName')['NewCases'].rolling(
    window = WINDOW, center=False).mean().fillna(0).reset_index(0, drop=True).round()
final_df['SmoothNewDeaths'] = final_df.groupby('CountryName')['NewDeaths'].rolling(
    window = WINDOW, center=False).mean().fillna(0).reset_index(0, drop=True).round()

# Compute percent change in new cases and deaths each day
final_df['CaseRatio'] = final_df.groupby('CountryName').SmoothNewCases.pct_change().fillna(0).replace(np.inf, 0) + 1
final_df['DeathRatio'] = final_df.groupby('CountryName').SmoothNewDeaths.pct_change().fillna(0).replace(np.inf, 0) + 1

# Add column for proportion of population infected
final_df['ProportionInfected'] = final_df['ConfirmedCases'] / final_df['Population2020']

# Create column of value to predict
final_df['PredictionRatio'] = final_df['CaseRatio'] / (1 - final_df['ProportionInfected'])

# Merge with vaccine
final_df = pd.merge(final_df, vaccine, how='left', on=['CountryName', 'Date'])
      
# Merge with weather
final_df = pd.merge(final_df, weather, how='left', on=['CountryName', 'Date'])

final_df.ConfirmedCases = final_df.ConfirmedCases.astype('int')
final_df.ConfirmedDeaths = final_df.ConfirmedDeaths.astype('int')

# View

final_df

# Save 

final_df.to_csv('../Data/Covid19_Europe_20210710_preprocessed.csv')
