==> Codes/17. Implementation.ipynb <==
# Packages

import numpy as np
import pandas as pd

%matplotlib inline
import matplotlib.pyplot as plt

from preprocessing import filter_country, colnames_generator, Create_TrainTest
from prediction import evaluate, onestep_prediction, multistep_prediction, plot
from prediction import train_lasso, train_ridge

import keras
import keras_tuner as kt

# Data 

data = pd.read_csv('../Data/Covid19_Europe_20210710_preprocessed.csv', index_col=0, parse_dates=['Date'])
data.info()

# Implementation

__Variants Emergence__

* Alpha $\rightarrow$ Sep-2020 
* Beta $\rightarrow$ May-2020 
* Gamma $\rightarrow$ Nov-2020 
* Delta $\rightarrow$ Oct-2020
* Lambda $\rightarrow$ Dec-2020
* Mu $\rightarrow$ Jan-2021

### 1. Extrcating data for each country

# Please add the necesary columns here

COLS = ['C1_School closing', 'C2_Workplace closing', 'C3_Cancel public events', 'C4_Restrictions on gatherings', 
        'C6_Stay at home requirements', 'H2_Testing policy', 'H3_Contact tracing', 'H6_Facial Coverings', 
        'H7_Vaccination policy'] 

T = 21       # days in past to include in training data
n_test = 30   # days to include for future

country_list = data.CountryName.sort_values().unique()

X_all_train_dict = {}
y_all_train_dict = {}
X_all_test_dict = {}
y_all_test_dict = {}


for country in country_list:
    
    X, y, indx = filter_country(data, 
                                country=country, 
                                COLS=COLS,
                                start_time='2019-01-01',   # Starting time
                                end_time='2020-10-01', 
                                outcome='PredictionRatio' )
    
    
    X_train, X_test, y_train, y_test, train_indx, test_indx = Create_TrainTest(X = X, 
                                                                               y = y,
                                                                               method='single', 
                                                                               indx=indx, 
                                                                               T=T, 
                                                                               n_test=n_test)
    
    X_all_train_dict[country] = X_train
    y_all_train_dict[country] = y_train
    
    X_all_test_dict[country] = X_test
    y_all_test_dict[country] = y_test

X_all_train_dict.keys()

train_indx[-1]

test_indx[-1]

### 2. Merging all countries

n_features = X.shape[1]

X_all_train = np.array([]).reshape(-1, n_features*T)
y_all_train = np.array([]).reshape(-1, 1)

X_all_test = np.array([]).reshape(-1, n_features*T)
y_all_test = np.array([]).reshape(-1, 1)


for country in country_list:
    X_all_train = np.concatenate((X_all_train, X_all_train_dict[country]))
    y_all_train = np.concatenate((y_all_train, y_all_train_dict[country]))
    
    X_all_test = np.concatenate((X_all_test, X_all_test_dict[country]))
    y_all_test= np.concatenate((y_all_test, y_all_test_dict[country]))

X_all_train.shape

# 3. Lasso 

lasso_trained_1step = train_lasso(X_all_train, 
                                  y_all_train, 
                                  scoring='neg_root_mean_squared_error', 
                                  n_cv=5)

### 3.1. Overall Performance 

lasso_pred_1step = lasso_trained_1step.predict(X_all_test)

pd.Series(evaluate(y_true=y_all_test, 
          y_pred=lasso_pred_1step), 
          name='all')

### 3.2. By Country 

lasso_1step_df = pd.DataFrame()

for country in country_list:
    temp = pd.Series(evaluate(y_true=y_all_test_dict[country], 
                              y_pred=lasso_trained_1step.predict(X_all_test_dict[country])), 
                     name=country)
    lasso_1step_df = pd.concat([lasso_1step_df, temp], axis='columns')

lasso_1step_df

# 4. Ridge 

ridge_trained_1step = train_ridge(X_all_train, 
                                  y_all_train, 
                                  scoring='neg_root_mean_squared_error', 
                                  n_cv=5)

### 4.1. Overall Performance

ridge_pred_1step = ridge_trained_1step.predict(X_all_test)

pd.Series(evaluate(y_true=y_all_test, 
          y_pred=ridge_pred_1step), 
          name='all')

### 4.2. By Country

ridge_1step_df = pd.DataFrame()

for country in country_list:
    temp = pd.Series(evaluate(y_true=y_all_test_dict[country], 
                              y_pred=ridge_trained_1step.predict(X_all_test_dict[country])), 
                     name=country)
    ridge_1step_df = pd.concat([ridge_1step_df, temp], axis='columns')

ridge_1step_df

# 5. ANN

### 5.1. Tuning Model

def model_builder(hp):
    
    # Initialize sequential API and start building model.
    model = keras.models.Sequential()
    model.add(keras.layers.Input(shape=(210)))
    
    # Tune the number of hidden layers and units in each.
    # Number of hidden layers: 1 - 5
    # Number of Units: 32 - 512 with stepsize of 32
    for i in range(1, hp.Int("num_layers", 1, 4)):
        model.add(
            keras.layers.Dense(
                units=hp.Int("units_" + str(i), min_value=20, max_value=200, step=20),
                activation="relu")
            )
        
        # Tune dropout layer with values from 0 - 0.3 with stepsize of 0.1.
        model.add(keras.layers.Dropout(hp.Float("dropout_" + str(i), 0, 0.3, step=0.1)))
    
    # Add output layer.
    model.add(keras.layers.Dense(1))
    
    # Tune learning rate for Adam optimizer with values from 0.01, 0.001, or 0.0001
    hp_learning_rate = hp.Choice("learning_rate", values=[1e-2, 1e-3, 1e-4])
    
    # Define optimizer, loss, and metrics
    model.compile(optimizer=keras.optimizers.adam_v2.Adam(learning_rate=hp_learning_rate),
                  loss='mse',
                  metrics=["mae", "mse"])
    
    return model

tuner = kt.Hyperband(hypermodel=model_builder,
                     objective = kt.Objective('val_mse', direction="min"),
                     max_epochs=100,
                     factor=3,
                     directory='MLP_hyperparameter',
                     project_name='Covid19')

stop_early = keras.callbacks.EarlyStopping(monitor='val_mse', patience=5)

tuner.search(X_all_train, 
             y_all_train, 
             epochs=100, 
             validation_data=(X_all_test, y_all_test), 
             callbacks=[stop_early], 
             verbose=2)

# Refit using the optimal hyperparameters
best_hps = tuner.get_best_hyperparameters()[0]
h_model = tuner.hypermodel.build(best_hps)

h_model.summary()

h_model.fit(X_all_train, 
            y_all_train, 
            epochs=100, 
            validation_data=(X_all_test, y_all_test), 
            callbacks=[stop_early], 
            verbose=2)

### 3.1. Overall Performance

mlp_pred_1step = h_model.predict(X_all_test)

pd.Series(evaluate(y_true=y_all_test, 
          y_pred=mlp_pred_1step), 
          name='all')

### By Country 

mlp_1step_df = pd.DataFrame()

for country in country_list:
    temp = pd.Series(evaluate(y_true=y_all_test_dict[country], 
                              y_pred=h_model.predict(X_all_test_dict[country])), 
                     name=country)
    mlp_1step_df = pd.concat([mlp_1step_df, temp], axis='columns')

mlp_1step_df

# LSTM

X_all_train_lstm = X_all_train.reshape(-1, T, n_features)
X_all_test_lstm = X_all_test.reshape(-1, T, n_features) 

X_all_train_lstm.shape

def model_builder_lstm(hp):
    
    # Initialize sequential API and start building model.
    model = keras.models.Sequential()
    model.add(keras.layers.Input(shape=(T, n_features)))
    
    # Tune the number of hidden layers and units in each.
    # Number of hidden layers: 1 - 5
    # Number of Units: 32 - 512 with stepsize of 32
    for i in range(1, hp.Int("num_layers", 1, 4)):
        model.add(
            keras.layers.LSTM(
                units=hp.Int("units_" + str(i), min_value=5, max_value=205, step=20),
                activation="tanh", return_sequences=True)
            )
        
        # Tune dropout layer with values from 0 - 0.3 with stepsize of 0.1.
        model.add(keras.layers.Dropout(hp.Float("dropout_" + str(i), 0, 0.5, step=0.1)))
    
    # Aggregating Layer
    model.add(keras.layers.GlobalMaxPool1D())
    
    # Add output layer.
    model.add(keras.layers.Dense(1))
    
    # Tune learning rate for Adam optimizer with values from 0.01, 0.001, or 0.0001
    hp_learning_rate = hp.Choice("learning_rate", values=[1e-2, 1e-3, 1e-4])
    
    # Define optimizer, loss, and metrics
    model.compile(optimizer=keras.optimizers.adam_v2.Adam(learning_rate=hp_learning_rate),
                  loss='mse',
                  metrics=["mae", "mse"])
    
    return model

tuner_lstm = kt.Hyperband(hypermodel=model_builder_lstm,
                         objective = kt.Objective('val_mse', direction="min"),
                         max_epochs=100,
                         factor=3,
                         directory='LSTM_hyperparameter',
                         project_name='Covid19')

stop_early = keras.callbacks.EarlyStopping(monitor='val_mse', patience=5)

tuner_lstm.search(X_all_train_lstm, 
                 y_all_train, 
                 epochs=100, 
                 validation_data=(X_all_test_lstm, y_all_test), 
                 callbacks=[stop_early], verbose=2)

best_hps_lstm = tuner_lstm.get_best_hyperparameters()[0]
h_lstm = tuner_lstm.hypermodel.build(best_hps_lstm)

h_lstm.summary()

best_hps_lstm = tuner_lstm.get_best_hyperparameters()[0]
h_lstm = tuner_lstm.hypermodel.build(best_hps_lstm)
h_lstm.fit(X_all_train_lstm, 
            y_all_train, 
            epochs=100, 
            validation_data=(X_all_test_lstm, y_all_test), 
            callbacks=[stop_early], 
            verbose=2)

### 3.1. Overall Performance

X_all_test_lstm[0:1].shape

h_lstm.summary()

h_lstm.predict(X_all_test_lstm[0:1])

lstm_pred_1step.shape

lstm_pred_1step = h_lstm.predict(X_all_test_lstm)

pd.Series(evaluate(y_true=y_all_test, 
          y_pred=lstm_pred_1step), 
          name='all')
